Rename RestoreBlockchains module
