+++ app/src/main/res/navigation/main_graph.xml
@@ -461,7 +461,7 @@
         android:name="io.horizontalsystems.bankwallet.modules.watchaddress.WatchAddressFragment" />
     <fragment
         android:id="@+id/restoreSelectCoinsFragment"
-        android:name="io.horizontalsystems.bankwallet.modules.restore.restoreselectcoins.RestoreSelectCoinsFragment"
+        android:name="io.horizontalsystems.bankwallet.modules.restore.restoreblockchains.RestoreBlockchainsFragment"
         tools:layout="@layout/fragment_manage_wallets" />
     <fragment
         android:id="@+id/receiveFragment"
